refactor(dsv4): align new module names with v0.25.1-dev conventions
Drop the model prefix from ops-layer file names (v0.25.1-dev mirrors
community vllm module names and never prefixes models in ops/), and
fix two collisions:

- ops/attention/deepseek_v4_cache.py      -> cache_utils.py (community name)
- ops/attention/deepseek_v4_sparse_index.py -> sparse_index.py
- ops/attention/dsv4_kv_insert_paths.py   -> kv_insert_paths.py
- ops/deepseek_v4_o_proj.py               -> o_proj.py
- patches/deepseek_v4.py                  -> dsv4.py (match dsv4_* family)
- runtime_utils.py                        -> adapter_utils.py (utils.py exists)

All imports, registry module paths and logger names updated; the
KUNLUN_DSV4_* env flags are unchanged.
