+++ vllm_kunlun/ops/attention/o_proj_alias.py
@@ -2,7 +2,7 @@
 
 DeepSeek-V4's NVIDIA submodules call this helper for the attention output
 projection; Kunlun has no deep-gemm backend, so the symbol is rebound to
-:func:`vllm_kunlun.ops.deepseek_v4_o_proj.deepseek_v4_bf16_o_proj`
+:func:`vllm_kunlun.ops.o_proj.deepseek_v4_bf16_o_proj`
 (inverse-RoPE + grouped BMM, with a PyTorch reference fallback).
 """
 import logging
@@ -10,7 +10,7 @@ import logging
 LOGGER = logging.getLogger("vllm_kunlun.ops.attention.o_proj_alias")
 _APPLIED_SENTINEL = "_dsv4_o_proj_wired"
 _O_PROJ_FN_NAME = "deep_gemm_fp8_o_proj"
-_EXPECTED_SOURCE = "vllm_kunlun.ops.deepseek_v4_o_proj"
+_EXPECTED_SOURCE = "vllm_kunlun.ops.o_proj"
 
 
 def _predicate(mod: object) -> bool:
@@ -21,7 +21,7 @@ def _predicate(mod: object) -> bool:
 
 
 def _applier(mod: object) -> None:
-    from ...ops.deepseek_v4_o_proj import deepseek_v4_bf16_o_proj
+    from ...ops.o_proj import deepseek_v4_bf16_o_proj
 
     original_fn = getattr(mod, _O_PROJ_FN_NAME, None)
     if original_fn is not None:
